# backend/tests/fixtures/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "Evidence": {
        "properties": {
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "quote": {
            "default": "",
            "title": "Quote",
            "type": "string"
          },
          "speaker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "title": "Evidence",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HardEscalation": {
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "ruleId": {
            "title": "Ruleid",
            "type": "string"
          },
          "sourceKnowledgeItemIds": {
            "items": {
              "type": "string"
            },
            "title": "Sourceknowledgeitemids",
            "type": "array"
          }
        },
        "required": [
          "ruleId",
          "reason"
        ],
        "title": "HardEscalation",
        "type": "object"
      },
...
      "KnowledgeItem": {
        "properties": {
          "confidence": {
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "dueDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedate"
          },
          "dueDateSourceText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duedatesourcetext"
          },
          "effectivePriority": {
            "anyOf": [
              {
                "enum": [
                  "CRITICAL",
                  "MAJOR",
                  "MINOR"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effectivepriority"
          },
          "evidence": {
            "$ref": "#/components/schemas/Evidence"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "manualOverride": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManualPriorityOverride"
              },
              {
                "type": "null"
              }
            ]
          },
          "meetingId": {
            "title": "Meetingid",
            "type": "string"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/Note"
            },
            "title": "Notes",
            "type": "array"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "relatedIds": {
            "items": {
              "type": "string"
            },
            "title": "Relatedids",
            "type": "array"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "stakeholders": {
            "items": {
              "type": "string"
            },
            "title": "Stakeholders",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          },
          "topicId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topicid"
          },
          "topicName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topicname"
          },
          "type": {
            "enum": [
              "IDEA",
              "DECISION",
              "ACTION",
              "QUESTION"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "description",
          "status",
          "confidence",
          "evidence",
          "meetingId"
        ],
        "title": "KnowledgeItem",
        "type": "object"
      },
      "ManualPriorityOverride": {
        "properties": {
          "overriddenAt": {
            "title": "Overriddenat",
            "type": "string"
          },
          "priority": {
            "enum": [
              "CRITICAL",
              "MAJOR",
              "MINOR"
            ],
            "title": "Priority",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "priority",
          "overriddenAt"
        ],
        "title": "ManualPriorityOverride",
        "type": "object"
      },
...
      "Note": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "body",
          "createdAt"
        ],
        "title": "Note",
        "type": "object"
      },
...
      "PriorityOverrideUpdate": {
        "properties": {
          "priority": {
            "anyOf": [
              {
                "enum": [
                  "CRITICAL",
                  "MAJOR",
                  "MINOR"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "PriorityOverrideUpdate",
        "type": "object"
      },
...
      "ReviewCandidate": {
        "properties": {
          "confidence": {
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "evidence": {
            "$ref": "#/components/schemas/Evidence"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "default": "PENDING",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED"
            ],
            "title": "Status",
            "type": "string"
          },
          "suggestedTopicId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestedtopicid"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "description",
          "reason",
          "confidence",
          "evidence"
        ],
        "title": "ReviewCandidate",
        "type": "object"
      },
      "ReviewStatusUpdate": {
        "properties": {
          "status": {
            "enum": [
              "ACCEPTED",
              "REJECTED"
            ],
            "title": "Status",
            "type": "string"
          },
          "topicId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topicid"
          }
        },
        "required": [
          "status"
        ],
        "title": "ReviewStatusUpdate",
        "type": "object"
      },
...
      "SemanticContribution": {
        "properties": {
          "contribution": {
            "title": "Contribution",
            "type": "number"
          },
          "disagreement": {
            "title": "Disagreement",
            "type": "boolean"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "scores": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Scores",
            "type": "object"
          }
        },
        "required": [
          "provider",
          "model",
          "scores",
          "contribution",
          "disagreement"
        ],
        "title": "SemanticContribution",
        "type": "object"
      },
...
      "Topic": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/TopicImage"
            },
            "title": "Images",
            "type": "array"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeItem"
            },
            "title": "Items",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/Note"
            },
            "title": "Notes",
            "type": "array"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TopicPriorityInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "stakeholders": {
            "items": {
              "type": "string"
            },
            "title": "Stakeholders",
            "type": "array"
          },
          "status": {
            "default": "Open",
            "enum": [
              "Open",
              "Closed"
            ],
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "items",
          "stakeholders"
        ],
        "title": "Topic",
        "type": "object"
      },
...
      "TopicImage": {
        "properties": {
          "contentType": {
            "title": "Contenttype",
            "type": "string"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "filename",
          "contentType",
          "size",
          "createdAt"
        ],
        "title": "TopicImage",
        "type": "object"
      },
...
      "TopicPriorityHistoryEntry": {
        "properties": {
          "algorithmVersion": {
            "title": "Algorithmversion",
            "type": "string"
          },
          "changedAt": {
            "title": "Changedat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "newPriority": {
            "enum": [
              "CRITICAL",
              "MAJOR",
              "MINOR"
            ],
            "title": "Newpriority",
            "type": "string"
          },
          "newScore": {
            "title": "Newscore",
            "type": "number"
          },
          "previousPriority": {
            "anyOf": [
              {
                "enum": [
                  "CRITICAL",
                  "MAJOR",
                  "MINOR"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previouspriority"
          },
          "previousScore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previousscore"
          },
          "primaryDrivers": {
            "items": {
              "type": "string"
            },
            "title": "Primarydrivers",
            "type": "array"
          },
          "sourceKnowledgeItemIds": {
            "items": {
              "type": "string"
            },
            "title": "Sourceknowledgeitemids",
            "type": "array"
          },
          "topicId": {
            "title": "Topicid",
            "type": "string"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "id",
          "topicId",
          "newPriority",
          "newScore",
          "changedAt",
          "algorithmVersion",
          "trigger"
        ],
        "title": "TopicPriorityHistoryEntry",
        "type": "object"
      },
      "TopicPriorityInfo": {
        "properties": {
          "algorithmVersion": {
            "title": "Algorithmversion",
            "type": "string"
          },
          "calculatedAt": {
            "title": "Calculatedat",
            "type": "string"
          },
          "calculatedPriority": {
            "enum": [
              "CRITICAL",
              "MAJOR",
              "MINOR"
            ],
            "title": "Calculatedpriority",
            "type": "string"
          },
          "calculatedScore": {
            "title": "Calculatedscore",
            "type": "number"
          },
          "confidence": {
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "effectivePriority": {
            "enum": [
              "CRITICAL",
              "MAJOR",
              "MINOR"
            ],
            "title": "Effectivepriority",
            "type": "string"
          },
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "hardEscalations": {
            "items": {
              "$ref": "#/components/schemas/HardEscalation"
            },
            "title": "Hardescalations",
            "type": "array"
          },
          "manualOverride": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManualPriorityOverride"
              },
              {
                "type": "null"
              }
            ]
          },
          "semanticContribution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SemanticContribution"
              },
              {
                "type": "null"
              }
            ]
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/TopicPrioritySignal"
            },
            "title": "Signals",
            "type": "array"
          }
        },
        "required": [
          "calculatedPriority",
          "calculatedScore",
          "effectivePriority",
          "confidence",
          "explanation",
          "calculatedAt",
          "algorithmVersion"
        ],
        "title": "TopicPriorityInfo",
        "type": "object"
      },
      "TopicPrioritySignal": {
        "properties": {
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "maxScore": {
            "title": "Maxscore",
            "type": "number"
          },
          "normalizedScore": {
            "title": "Normalizedscore",
            "type": "number"
          },
          "rawValue": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rawvalue"
          },
          "sourceKnowledgeItemIds": {
            "items": {
              "type": "string"
            },
            "title": "Sourceknowledgeitemids",
            "type": "array"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "weightedScore": {
            "title": "Weightedscore",
            "type": "number"
          }
        },
        "required": [
          "type",
          "normalizedScore",
          "weightedScore",
          "maxScore",
          "explanation"
        ],
        "title": "TopicPrioritySignal",
        "type": "object"
      },
...
      "TopicProposalAccept": {
        "properties": {
          "existingTopicId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Existingtopicid"
          },
          "suggestedName": {
            "title": "Suggestedname",
            "type": "string"
          },
          "topicName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topicname"
          }
        },
        "required": [
          "suggestedName"
        ],
        "title": "TopicProposalAccept",
        "type": "object"
      },
      "TopicProposalReject": {
        "properties": {
          "suggestedName": {
            "title": "Suggestedname",
            "type": "string"
          }
        },
        "required": [
          "suggestedName"
        ],
        "title": "TopicProposalReject",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "second-brain backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/priority-history/recent": {
      "get": {
        "operationId": "get_recent_priority_escalations_api_priority_history_recent_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 14,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TopicPriorityHistoryEntry"
                  },
                  "title": "Response Get Recent Priority Escalations Api Priority History Recent Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recent Priority Escalations"
      }
    },
...
    "/api/review/topic-proposals/accept": {
      "post": {
        "operationId": "accept_topic_proposal_api_review_topic_proposals_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopicProposalAccept"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeItem"
                  },
                  "title": "Response Accept Topic Proposal Api Review Topic Proposals Accept Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Topic Proposal"
      }
    },
    "/api/review/topic-proposals/reject": {
      "post": {
        "operationId": "reject_topic_proposal_api_review_topic_proposals_reject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopicProposalReject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Topic Proposal"
      }
    },
    "/api/review/{candidate_id}": {
      "patch": {
        "operationId": "update_review_status_api_review__candidate_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCandidate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Review Status"
      }
    },
...
    "/api/topics/{topic_id}/priority-history": {
      "get": {
        "operationId": "get_topic_priority_history_api_topics__topic_id__priority_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "topic_id",
            "required": true,
            "schema": {
              "title": "Topic Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TopicPriorityHistoryEntry"
                  },
                  "title": "Response Get Topic Priority History Api Topics  Topic Id  Priority History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Topic Priority History"
      }
    },
    "/api/topics/{topic_id}/priority-override": {
      "patch": {
        "operationId": "set_topic_priority_override_api_topics__topic_id__priority_override_patch",
        "parameters": [
          {
            "in": "path",
            "name": "topic_id",
            "required": true,
            "schema": {
              "title": "Topic Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriorityOverrideUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Topic Priority Override"
      }
    },
    "/api/topics/{topic_id}/recalculate-priority": {
      "post": {
        "operationId": "recalculate_topic_priority_api_topics__topic_id__recalculate_priority_post",
        "parameters": [
          {
            "in": "path",
            "name": "topic_id",
            "required": true,
            "schema": {
              "title": "Topic Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Topic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recalculate Topic Priority"
      }
    },
...
  }
}